changed dbtest project organization
- split the code from a single main.cpp file to 5 files in c++ spirit
(each class has a header and an implementation file)
- commented out the run 1000 times code (annoying to touch the screen by
mistake and have the test run 1000 times)
