+++ examples/cpp/DatabaseTest/.mosyncproject
@@ -1,10 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project supports-build-configs="true" version="1.4">
+<project supports-build-configs="true" version="1.5">
 <build.cfg id="Debug" types="Debug"/>
 <build.cfg id="Release" types="Release"/>
+<criteria>
+<filter type="com.mobilesorcery.sdk.capabilities.devices.elementfactory">
+<capabilities optional="" required=""/>
+</filter>
+</criteria>
 <properties>
 <property key="build.prefs:additional.libraries" value="MAUtil.lib"/>
 <property key="build.prefs:additional.libraries/Debug" value="MAUtilD.lib"/>
+<property key="build.prefs:gcc.warnings/Debug" value="10"/>
+<property key="build.prefs:gcc.warnings/Release" value="10"/>
+<property key="profile.mgr.type" value="0"/>
 <property key="template.id" value="project.moblet"/>
 </properties>
 </project>
